Rename mujoco_model_gen to mujoco_model_generator
Package names should follow class names.
