+++ rllab/envs/mujoco/dynamics_randomization/__init__.py
@@ -1,5 +1,5 @@
 from rllab.envs.mujoco.dynamics_randomization.variation import Variations
 from rllab.envs.mujoco.dynamics_randomization.variation import Method
 from rllab.envs.mujoco.dynamics_randomization.variation import Distribution
-from rllab.envs.mujoco.dynamics_randomization.mujoco_model_gen import MujocoModelGenerator
+from rllab.envs.mujoco.dynamics_randomization.mujoco_model_generator import MujocoModelGenerator
 from rllab.envs.mujoco.dynamics_randomization.randomized_env import randomize
